pyproject.toml
[project]
name = "banking-agent"
version = "0.1.0"
description = "Banking Agent with MCP Server and Multi-LLM Support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "fastapi>=0.115.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server", "mcp_client"]

